Fix compiler warnings in cyxwiz-backend
- Fix signed/unsigned conversion warnings (C4244, C4267)
  - Cast dim_t to int for array indexing in loss.cpp, layer.cpp
  - Cast size_t to unsigned int for af::tile in linear.cpp, dimensionality_reduction.cpp
  - Cast dim_t to double for af::seq in layer.cpp
- Fix unreferenced local variable warnings (C4189)
  - Add (void) casts for unused variables in clustering.cpp, linear.cpp, layer.cpp
- Fix deprecated API warning (C4996)
  - Update af::stdev to use AF_VARIANCE_SAMPLE parameter in data_transform.cpp

+++ cyxwiz-backend/src/algorithms/dimensionality_reduction.cpp
@@ -400,7 +400,7 @@ std::vector<std::vector<double>> DimensionalityReduction::ComputeSquaredDistance
             af::array dot_products = af::matmul(X, af::transpose(X));  // [n x n]
 
             // Broadcast sq_norms to [n x n]
-            af::array dist_sq = af::tile(sq_norms, 1, n) + af::tile(af::transpose(sq_norms), n, 1) - 2.0 * dot_products;
+            af::array dist_sq = af::tile(sq_norms, 1, static_cast<unsigned int>(n)) + af::tile(af::transpose(sq_norms), static_cast<unsigned int>(n), 1) - 2.0 * dot_products;
             dist_sq = af::max(dist_sq, 0.0);  // Clamp negative values due to numerical errors
 
             // Copy back to CPU
